arch: move x86 base code from pm.c to x86/pm.c
