+++ arch/x86/Makefile
@@ -1,4 +1,4 @@
-OBJS = cpu.o ide.o
+OBJS = cpu.o ide.o pm.o
 
 %.o: %.c
 	$(CC) $(CFLAGS) -c -I $(BUILDROOT)/include $<
